server-ai/requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
pandas
google-genai
httpx
